add pause detection for gpx files from GPS Master
This fix is so far only activated for gpx files created by 'GPS Master'
(e.g. for Ultrasport NavRun) but can easily be activated for all gpx
files. (It does not look for our configuration flag so far)

This fix looks for the average recording interval and will strip larger
gaps (currently: factor 10) as pauses.