add gradient to elevation plot
There is a new class to calculate gradient series for routes with the option to set a specific moving average (i.e. which kernel, which kernel width). Currently, this series is not calculated automatically (as pace is for trackdata). So far, this series is only shown in tooltips for plots containing the elevation (elevation/pace+hr+elevation/power).

Feel free to discuss the current choice (uniform, width = 20).