add configuration manager as service
This adds a new configuration component, coming with a general
configuration list that contains key-value-pairs only and can provide
access to detailed category objects. The configuration is available
through a configuration manager service that serves the configuration
for any account and comes with a twig extension.

Example usage is shown in vdot analysis tool.